Parent spawned clusters to layer entity

+++ Runtime/Components/CGeometryCluster.cs
@@ -5,7 +5,7 @@ namespace henningboat.CubeMarching.Runtime.Components
 {
     public struct CGeometryCluster:IComponentData
     {
-        public int3 PositionWS;
+        public float3 PositionWS;
         public Entity LayerEntity;
     }
 }
